+++ CHANGELOG.md
@@ -9,6 +9,9 @@ New entries in this file should aim to provide a meaningful amount of informatio
 
 ## [Unreleased]
 
+### Added
+* read_only behavior to items and thesis [[#3758](https://github.com/ualbertalib/jupiter/issues/3758)]
+
 ## 2.10.3 - 2025-03-06
 
 ### Changed