+++ CHANGELOG.md
@@ -9,6 +9,9 @@ New entries in this file should aim to provide a meaningful amount of informatio
 
 ## [Unreleased]
 
+### Changed
+* pin chrome so ci passes [3764](https://github.com/ualbertalib/jupiter/issues/3764)
+
 ## 2.10.3 - 2025-03-06
 
 ### Changed